@@ -2,7 +2,6 @@ import 'dart:async';
 import 'package:flutter/services.dart';
 import 'package:meta/meta.dart' show visibleForTesting;
 
-
 typedef DetectPluggedCallback = Function(HeadsetState payload);
 
 enum HeadsetState {
@@ -13,20 +12,19 @@ enum HeadsetState {
 }
 
 class HeadsetEvent {
-  
-  static HeadsetEvent _instance;
+  static HeadsetEvent? _instance;
   final MethodChannel _channel;
-  DetectPluggedCallback detectPluggedCallback;
+  DetectPluggedCallback? detectPluggedCallback;
 
-  factory HeadsetEvent() {    
+  factory HeadsetEvent() {
     if (_instance == null) {
       final MethodChannel methodChannel =
-          const MethodChannel('flutter.moum/headset_event');
+      const MethodChannel('flutter.moum/headset_event');
       _instance = HeadsetEvent.private(methodChannel);
     }
-    return _instance;
+    return _instance!;
   }
-  
+
   @visibleForTesting
   HeadsetEvent.private(this._channel);
 
@@ -37,7 +35,7 @@ class HeadsetEvent {
 
   Future<HeadsetState> get getCurrentState async {
     final int state = await _channel.invokeMethod('getCurrentState');
-    switch(state) {
+    switch (state) {
       case 0:
         return Future.value(HeadsetState.DISCONNECT);
       case 1:
@@ -54,20 +52,21 @@ class HeadsetEvent {
   }
 
   Future<dynamic> _handleMethod(MethodCall call) async {
-    switch(call.method) {
-      case "connect":
-        return detectPluggedCallback(HeadsetState.CONNECT);
-      case "disconnect":
-        return detectPluggedCallback(HeadsetState.DISCONNECT);
-      case "nextButton":
-        return detectPluggedCallback(HeadsetState.NEXT);
-      case "prevButton":
-        return detectPluggedCallback(HeadsetState.PREV);
-      default:
-        print('No idea');
-        return detectPluggedCallback(HeadsetState.DISCONNECT);
+    DetectPluggedCallback? detectPluggedCallbackCopy = detectPluggedCallback;
+    if (detectPluggedCallbackCopy != null) {
+      switch (call.method) {
+        case "connect":
+          return detectPluggedCallbackCopy(HeadsetState.CONNECT);
+        case "disconnect":
+          return detectPluggedCallbackCopy(HeadsetState.DISCONNECT);
+        case "nextButton":
+          return detectPluggedCallbackCopy(HeadsetState.NEXT);
+        case "prevButton":
+          return detectPluggedCallbackCopy(HeadsetState.PREV);
+        default:
+          print('No idea');
+          return detectPluggedCallbackCopy(HeadsetState.DISCONNECT);
+      }
     }
   }
-
-
 }
